Computer_vision_medical_images/requirements.txt
Pillow
numpy
scikit-image
Shapely
